+++ packages/core/CHANGELOG.md
@@ -1,3 +1,5 @@
+## 0.0.24
+- BUGFIX: Prevent button w/href height from expanding due to IE fix
 ## 0.0.23
 - never use esm for babel runtime
 ## 0.0.22