+++ packages/mobile-components/CHANGELOG.md
@@ -1,3 +1,5 @@
+## 0.0.27
+- BUGFIX: Prevent button w/href height from expanding due to IE fix
 ## 0.0.26
 - never use esm for babel runtime
 ## 0.0.25